requirements.txt
aiohttp==3.9.5
aiosqlite==0.20.0
discord.py==2.3.2
discord_ios==1.1.1
python-dotenv==1.0.1
pytz==2024.1
datetime

